move EliminateRedundantTransitions into separate file
